requirements.txt
# PrizePicks & Underdog EV Bot + API Dependencies

# Discord Bot
discord.py>=2.0.0

# Web Framework (API)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# HTTP Client
aiohttp>=3.9.0

# Environment
python-dotenv>=1.0.0

# Fuzzy Matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
